drm/bridge: synopsys: dw-hdmi-qp: Make sure video data path is enable when start transmit video
When conducting FRL training, the video data transmitted from
VOP to HDMI must be disabled. Until the training is successful
or fails, then it will be reopened. When the FRL training fails,
the video transmission from the video to HDMI is not re-enabled.
This will result in the inability to display normally even when
switching to tmds mode after training fails.
Therefore, regardless of the specific scenario, when switching
to the tmds mode, the data transmission from VOP to HDMI must be
enabled.

Change-Id: I1c853c8197fdf7aaef32c80c5cdb73db13d9ec4c
Signed-off-by: Algea Cao <algea.cao@rock-chips.com>